+++ data/data_8107224.s
@@ -1,51 +1,7 @@
 .section .rodata
 
 
-.align 2
-.global gUnknown_8107374
-gUnknown_8107374: @ 8107374
-.byte 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 
-.global gUnknown_8107380
-gUnknown_8107380: @ 8107380
-.byte 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-@ ???
-.string "pksdir0\0"
-@ code_8086A3C #1
-.string "pksdir0\0"
-
-.global gUnknown_810739C
-gUnknown_810739C: @ 810739C
-.byte 0x04, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x03, 0x00, 0x02, 0x00, 0x18, 0x00, 0x00, 0x00
-.byte 0x04, 0x00, 0x02, 0x00, 0x18, 0x00, 0x00, 0x00, 0x05, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00
-.byte 0x06, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-.byte 0x00, 0x00, 0x0b, 0x00, 0x08, 0x00, 0x00, 0x00
-
-.global gUnknown_81073D4
-gUnknown_81073D4: @ 81073D4
-.byte 0x04, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x03, 0x00, 0x02, 0x00, 0x18, 0x00, 0x00, 0x00
-.byte 0x04, 0x00, 0x02, 0x00, 0x18, 0x00, 0x00, 0x00, 0x05, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00
-.byte 0x06, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-.byte 0x00, 0x00, 0x0c, 0x00, 0x08, 0x00, 0x00, 0x00
-
-.global gUnknown_810740C
-gUnknown_810740C: @ 810740C
-.byte 0x04, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x05, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00
-.byte 0x04, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x04, 0x00, 0x02, 0x00, 0x18, 0x00, 0x00, 0x00
-.byte 0x03, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-.byte 0x00, 0x00, 0x0a, 0x00, 0x08, 0x00, 0x00, 0x00
-@ code_8086A3C #2
-.string "pksdir0\0"
-
-.global gUnknown_810744C
-gUnknown_810744C: @ 810744C
-.byte 0x01, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-.byte 0x00, 0x00, 0x09, 0x00, 0x09, 0x00, 0x00, 0x00
-
-.global gUnknown_8107464
-gUnknown_8107464: @ 8107464
-.byte 0x05, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x04, 0x00, 0x02, 0x00, 0x0c, 0x00, 0x00, 0x00
-.byte 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x07, 0x00, 0x00, 0x00
 .string "pksdir0\0"
 .string "pksdir0\0"
 .string "pksdir0\0"
