+++ ld_script.txt
@@ -282,6 +282,9 @@ SECTIONS {
         src/dungeon_cleared_window.o(.text);
         src/dungeon_jobs.o(.text);
         src/code_80848F0.o(.text);
+        src/dungeon_cutscene_skarmory.o(.text);
+        src/dungeon_cutscene_meanies.o(.text);
+        src/dungeon_cutscene_zapdos.o(.text);
         src/dungeon_cutscenes.o(.text);
         src/pokemon.o(.text);
         src/pokemon_3.o(.text);
@@ -622,6 +625,9 @@ SECTIONS {
         src/dungeon_cleared_window.o(.rodata);
         src/dungeon_jobs.o(.rodata);
         src/code_80848F0.o(.rodata);
+        src/dungeon_cutscene_skarmory.o(.rodata);
+        src/dungeon_cutscene_meanies.o(.rodata);
+        src/dungeon_cutscene_zapdos.o(.rodata);
         data/data_8107224.o(.rodata);
         src/pokemon.o(.rodata);
         data/data_81076E4.o(.rodata);
